refactor(anyharness): split adapter operations by capability
