+++ anyharness/crates/anyharness-lib/src/adapters/files/mod.rs
@@ -1,3 +1,7 @@
+pub mod operations;
 pub mod safety;
 pub mod service;
 pub mod types;
+
+#[cfg(test)]
+mod service_tests;
