+++ anyharness/crates/anyharness-lib/src/adapters/files/operations/mod.rs
@@ -1,0 +1,8 @@
+pub mod create;
+pub mod delete;
+mod entry;
+pub mod list;
+pub mod read;
+pub mod rename;
+pub mod stat;
+pub mod write;
