+++ anyharness/crates/anyharness-lib/src/adapters/git/operations/staging.rs
@@ -1,0 +1,33 @@
+use std::path::{Path, PathBuf};
+
+use super::super::executor::run_git_ok;
+
+pub fn stage_paths(workspace_path: &Path, paths: &[String]) -> anyhow::Result<()> {
+    if paths.is_empty() {
+        return Ok(());
+    }
+    let repo_root = run_git_ok(workspace_path, &["rev-parse", "--show-toplevel"])?
+        .trim()
+        .to_string();
+    let repo_root_path = PathBuf::from(&repo_root);
+    let mut args = vec!["add", "--"];
+    let path_refs: Vec<&str> = paths.iter().map(|s| s.as_str()).collect();
+    args.extend(path_refs);
+    run_git_ok(&repo_root_path, &args)?;
+    Ok(())
+}
+
+pub fn unstage_paths(workspace_path: &Path, paths: &[String]) -> anyhow::Result<()> {
+    if paths.is_empty() {
+        return Ok(());
+    }
+    let repo_root = run_git_ok(workspace_path, &["rev-parse", "--show-toplevel"])?
+        .trim()
+        .to_string();
+    let repo_root_path = PathBuf::from(&repo_root);
+    let mut args = vec!["reset", "HEAD", "--"];
+    let path_refs: Vec<&str> = paths.iter().map(|s| s.as_str()).collect();
+    args.extend(path_refs);
+    run_git_ok(&repo_root_path, &args)?;
+    Ok(())
+}
