+++ anyharness/crates/anyharness-lib/src/adapters/hosting/mod.rs
@@ -1,4 +1,5 @@
 pub mod github_cli;
+pub mod operations;
 pub mod service;
 pub mod types;
