+++ anyharness/crates/anyharness-lib/src/adapters/hosting/operations/create_pr.rs
@@ -1,0 +1,33 @@
+use std::path::Path;
+
+use super::super::github_cli;
+use super::super::types::{CreatePullRequestResult, HostingServiceError};
+use crate::adapters::git::GitService;
+
+pub fn create_pull_request(
+    workspace_path: &Path,
+    title: &str,
+    body: Option<&str>,
+    base_branch: &str,
+    draft: bool,
+) -> Result<CreatePullRequestResult, HostingServiceError> {
+    let repo_root = GitService::resolve_repo_root(workspace_path)
+        .map_err(|error| HostingServiceError::PullRequestCreateFailed(error.to_string()))?;
+
+    match github_cli::create_pr(&repo_root, title, body, base_branch, draft) {
+        Ok(pr) => Ok(CreatePullRequestResult {
+            pull_request: pr,
+            manual_url: None,
+        }),
+        Err(github_cli::GhError::NotInstalled) => Err(HostingServiceError::GhNotInstalled),
+        Err(github_cli::GhError::AuthRequired(msg)) => {
+            Err(HostingServiceError::GhAuthRequired(msg))
+        }
+        Err(github_cli::GhError::NoPrFound) => Err(HostingServiceError::PullRequestCreateFailed(
+            "PR was created but could not be read back".into(),
+        )),
+        Err(github_cli::GhError::CommandFailed(msg)) => {
+            Err(HostingServiceError::PullRequestCreateFailed(msg))
+        }
+    }
+}
